Create Stack navigators for each tab.
The code is very similar, only difference is the ParamList used. I tried to get this done in the buildStack function we had in App.tsx but I couldn't come up with a clean way to do it.
If you have an idea to get rid of this repeated code suggest it.